% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishhook.R
\name{aggregate.targets}
\alias{aggregate.targets}
\title{title}
\usage{
aggregate.targets(targets, by = NULL, fields = NULL, rolling = NULL,
  disjoint = TRUE, na.rm = FALSE, FUN = list(), verbose = TRUE)
}
\arguments{
\item{targets}{annotated GRanges of targets with fields $coverage, optional field, $count and additional numeric covariates, or path to .rds file of the same; path to bed or rds containing genomic target regions with optional target name}

\item{by}{character vector with which to split into meta-territories (default == NULL)}

\item{fields}{by default all meta data fields of targets EXCEPT reserved field names $coverage, $counts, $query.id (default == NULL)}

\item{rolling}{if specified, positive integer specifying how many (genome coordinate) adjacent to aggregate in a rolling fashion; positive integer with which to performa rolling sum / weighted average WITHIN chromosomes of "rolling" ranges" --> return a granges (default == NULL)}

\item{disjoint}{boolean only take disjoint bins of input (default == TRUE)}

\item{na.rm}{boolean only applicable for sample wise aggregation (i.e. if by = NULL) (default == FALSE)}

\item{FUN}{list only applies (for now) if by = NULL, this is a named list of functions, where each item named "nm" corresponds to an optional function of how to alternatively aggregate field "nm" per samples, for alternative aggregation of coverage and count.  This function is applied at every iteration of loading a new sample and adding to the existing set.   It is normally sum [for coverage and count] and coverage weighted mean [for all other covariates].  Alternative coverage / count aggregation functions should have two arguments (val1, val2) and all other alt covariate aggregation functions should have four arguments (val1, cov1, val2, cov2) where val1 is the accumulating vector and val2 is the new vector of values.}

\item{verbose}{boolean verbose flag (default == TRUE)}
}
\value{
GRangesList of input targets annotated with new aggregate covariate statistics OR GRanges if rolling is specified
}
\description{
Gathers annotated targets across a vector "by" into meta-intervals returned as a GRangesList, and returns the
aggregated statistics for these meta intervals by summing coverage and counts, and performing a weighted average of all other meta data fields
(except query.id)

If rolling = TRUE, will return a rolling collapse of the sorted input where "rolling" specifies the number of adjacent intervals that are aggregated in a rolling manner.
(only makes sense for tiled target sets)

If by = NULL and targets is a vector of path names, then aggregation will be done "sample wise" on the files, ie each .rds input will be assumed to comprise the same 
intervals in teh same order and aggregation will be computed coverage-weighted mean of covariates, a sum of coverage and counts, and (if present) a Fisher combined
of $p values.  Covariates are inferred from the first file in the list.
}
\author{
Marcin Imielinski
}
