% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishhook.R
\name{annotate.targets}
\alias{annotate.targets}
\title{title}
\usage{
annotate.targets(targets, covered = NULL, events = NULL, mc.cores = 1,
  na.rm = TRUE, pad = 0, verbose = TRUE, max.slice = 1000,
  ff.chunk = 1e+06, max.chunk = 1e+11, out.path = NULL,
  covariates = list(), maxpatientpergene = Inf, ptidcol = NULL,
  weightEvents = FALSE, ...)
}
\arguments{
\item{targets}{path to bed or rds containing genomic target regions with optional target name}

\item{covered}{optional path to bed or rds containing  granges object containing "covered" genomic regions (default = NULL)}

\item{events}{optional path to bed or rds containing ranges corresponding to events (ie mutations etc) (default = NULL)}

\item{mc.cores}{integer info (default = 1)}

\item{na.rm}{info (default = TRUE)}

\item{pad}{info (default = 0)}

\item{verbose}{boolean verbose flag (default = FALSE)}

\item{max.slice}{integer Max slice of intervals to evaluate with  gr.val (default = 1e3)}

\item{ff.chunk}{integer Max chunk to evaluate with fftab (default = 1e6)}

\item{max.chunk}{integer gr.findoverlaps parameter (default = 1e11)}

\item{out.path}{string Pathname to output file (default = NULL)}

\item{covariates}{list}

\item{maxpatientpergene}{Sets the maximum number of events a patient can contribute per target (default = Inf)}

\item{ptidcol}{string Column where patient ID is stored}

\item{...}{paths to sequence covariates whose output names will be their argument names, and each consists of a list with (default = FALSE)
$track field corresponding to a GRanges, RleList, ffTrack object (or path to rds containing that object), $type which can
have one of three values "numeric", "sequence", "interval".
Numeric tracks must have $score field if they are GRanges), and can have a $na.rm logical field describing how to treat NA values
(set to na.rm argument by default)
Sequence covariates must be ffTrack objects (or paths to ffTrack rds) and require an additional variables $signatures, which
will be used as input to fftab, and can have optional logical argument $grep to specify inexact matches (see fftab)
Interval covariates must be Granges (or paths to GRanges rds) or paths to bed files}

\item{weightEvetns}{boolean If TRUE, will weight events by their overlap with targets. e.g. if 10% of an event overlaps with a target 
region, that target region will get assigned a score of 0.1 for that event. If false, any overlap will be given a weight of 1.}
}
\value{
GRanges of input targets annotated with covariate statistics (+/- constrained to the subranges in optional argument covered)
}
\description{
Takes input of GRanges targets, an optional set of "covered" intervals, and an indefinite list of covariates which can be R objects
(GRanges, ffTrack, Rle) or file paths to .rds, .bw, .bed files, and an annotated target intervals GRanges with covariates computed
for each interval.   These target intervals can be further annotated with mutation counts and plugged into a generalized linear regression
(or other) model downstream.

There are three types of covariates: numeric, sequence, interval.  The covariates are computed as follows:
numeric covariates: the mean value
sequence covarites: fraction of bases satisfying $signature
interval covariates: fraction of bases overlapping feature
}
\author{
Marcin Imielinski
}
