% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishhook.R
\name{c.Cov}
\alias{c.Cov}
\title{title}
\usage{
\method{c}{Cov}(...)
}
\arguments{
\item{...}{A series of Cov or Cov_Arr (Covariates), note all objects must be of type Cov or Cov_Arr}
}
\value{
Cov_Arr object that can be passed directly into the FishHook object constructor
}
\description{
Override the c operator for covariates so that when you type: c(Cov1,Cov2,Cov3) it returns a Cov_Arr object that support vector like operation.
}
\author{
Zoran Z. Gajic
}
