% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishhook.R
\docType{data}
\name{Cov}
\alias{Cov}
\title{Cov}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Cov
}
\arguments{
\item{Covariate}{object of type, GRanges, ffTrack, RleList or character. Note that character objects must be paths to files containing one of the other types as a .rds file}

\item{type}{a string indicating the type of Covariate, valid options are: numeric, sequence, interval. See Annotate Targets for more information on Covariate types}

\item{signature}{In the case where a ffTrack object is of type sequence, a signature field is required, see fftab in ffTrack for more information.}
}
\value{
Cov object that can be passed to FishHook object constructor
}
\description{
Stores Covariate for passing to FishHook object. To be packaged in the Cov_Array Class by calling c(Cov1,Cov2,Cov3)
}
\author{
Zoran Z. Gajic
}
\keyword{datasets}
