% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishhook.R
\name{length.Cov_Arr}
\alias{length.Cov_Arr}
\title{title}
\usage{
\method{length}{Cov_Arr}(obj, ...)
}
\arguments{
\item{obj}{Cov_Arr object that is passed to the length function}
}
\value{
number of covariates contained in the Cov_Arr object as defined by length(Cov_Arr$cvs)
}
\description{
Overrides the length function 'length(Cov_Arr)' for use with Cov_Arr
}
\author{
Zoran Z. Gajic
}
