% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishhook.R
\name{length.FishHook}
\alias{length.FishHook}
\title{title}
\usage{
\method{length}{FishHook}(obj, ...)
}
\arguments{
\item{obj}{FishHook object that is passed to the length function}
}
\value{
length of the targets contained in the FishHook object
}
\description{
Overrides the length function 'length(FishHook)' for use with FishHook
}
\author{
Zoran Z. Gajic
}
