% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishhook.R
\docType{data}
\name{Score}
\alias{Score}
\title{title}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Score
}
\arguments{
\item{annotated}{The annotated targets as an output from Annotated$scoreTargets() or the standard score.targets().}
}
\value{
Score object that can be plotted/analyzed
}
\description{
Stores the scored targets. Note that this constructors should be called from Annotated$scoreTargets(). Scores can also be plotted on qqplots using included functions. For other params see score.targets()
}
\author{
Zoran Z. Gajic
}
\keyword{datasets}
