% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fishhook.R
\name{score.targets}
\alias{score.targets}
\title{score.targets}
\usage{
score.targets(targets, covariates = names(values(targets)), model = NULL,
  return.model = FALSE, nb = TRUE, verbose = TRUE, iter = 200,
  subsample = 1e+05, seed = NULL, p.randomized = TRUE)
}
\arguments{
\item{targets}{annotated targets with fields $coverage, optional field, $count and additional numeric covariates}
}
\value{
GRanges of scored results
}
\description{
Scores targets based on covariates using gamma-poisson model with coverage as constant
}
\author{
Marcin Imielinski
}

