% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishhook.R
\name{[.Annotate}
\alias{[.Annotate}
\title{[.Annotate}
\usage{
\method{[}{Annotate}(obj, range)
}
\arguments{
\item{obj}{This is the Annotated object to be subset}

\item{range}{This is the range of targets to return, like subsetting a vector. e.g. c(1,2,3,4,5)[3:4] == c(3,4)}
}
\value{
Annotated object that can manipulated and scored, but cannot be aggregated again.
}
\description{
Overrides the "[" operator for the Annotated object. This allows subsetting of the annotated data in Annotated Objects.
}
\author{
Zoran Z. Gajic
}
