% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishhook.R
\name{[.Cov_Arr}
\alias{[.Cov_Arr}
\title{[.Cov_Arr}
\usage{
\method{[}{Cov_Arr}(obj, range)
}
\arguments{
\item{obj}{This is the Cov_Arr to be subset}

\item{range}{This is the range of Covariates to return, like subsetting a vector. e.g. c(1,2,3,4,5)[3:4] == c(3,4)}
}
\value{
A new Cov_Arr object that contains only the Covs within the given range
}
\description{
Overrides the subset operator x[] for use with Cov_Arr to allow for vector like subsetting
}
\author{
Zoran Z. Gajic
}
