% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishhook.R
\name{[.FishHook}
\alias{[.FishHook}
\title{title}
\usage{
\method{[}{FishHook}(obj, i = NULL, j = NULL)
}
\arguments{
\item{obj}{FishHook object This is the FishHookObject to be subset}

\item{i}{vector subset hypotheses}

\item{j}{vector subset covariates}
}
\value{
A new FishHook object that contains only the given hypotheses and/or covariates
}
\description{
Overrides the subset operator x[] for use with FishHook to allow for vector like subsetting, see fishHook demo for examples
}
\author{
Zoran Z. Gajic
}
