% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishhook.R
\name{aggregate.targets}
\alias{aggregate.targets}
\title{aggregate.targets}
\usage{
aggregate.targets(targets, by = NULL, fields = NULL, rolling = NULL,
  disjoint = TRUE, na.rm = FALSE, FUN = list(), verbose = TRUE)
}
\arguments{
\item{targets}{annotated GRanges of targets with fields $coverage, optional field, $count and additional numeric covariates, or path to .rds file of the same}

\item{by}{character vector with which to split into meta-territories}

\item{fields}{by default all meta data fields of targets EXCEPT reserved field names $coverage, $counts, $query.id}

\item{rolling}{if specified, positive integer specifying how many (genome coordinate) adjacent to aggregate in a rolling fashion}
}
\value{
GRangesList of input targets annotated with new aggregate covariate statistics OR GRanges if rolling is specified
}
\description{
Gathers annotated targets across a vector "by" into meta-intervals returned as a GRangesList, and returns the
aggregated statistics for these meta intervals by summing coverage and counts, and performing a weighted average of all other meta data fields
(except query.id)
}
\details{
If rolling = TRUE, will return a rolling collapse of the sorted input where "rolling" specifies the number of adjacent intervals that are aggregated in a rolling manner.
(only makes sense for tiled target sets)

If by = NULL and targets is a vector of path names, then aggregation will be done "sample wise" on the files, ie each .rds input will be assumed to comprise the same 
intervals in teh same order and aggregation will be computed coverage-weighted mean of covariates, a sum of coverage and counts, and (if present) a Fisher combined
of $p values.  Covariates are inferred from the first file in the list.
}
\author{
Marcin Imielinski
}
