% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishhook.R
\name{annotate.targets}
\alias{annotate.targets}
\title{annotate.targets}
\usage{
annotate.targets(targets, covered = NULL, events = NULL, ...,
  mc.cores = 1, na.rm = TRUE, pad = 0, verbose = TRUE,
  max.slice = 1000, ff.chunk = 1e+06, max.chunk = 1e+11,
  out.path = NULL, covariates = list(), maxPtGene = NULL)
}
\arguments{
\item{targets}{path to bed or rds containing genomic target regions with optional target name}

\item{covered}{optional path to bed or rds containing  granges object containing "covered" genomic regions}

\item{events}{optional path to bed or rds containing ranges corresponding to events (ie mutations etc)}

\item{...}{paths to sequence covariates whose output names will be their argument names, and each consists of a list with
$track field corresponding to a GRanges, RleList, ffTrack object (or path to rds containing that object), $type which can
have one of three values "numeric", "sequence", "interval".
Numeric tracks must have $score field if they are GRanges), and can have a $na.rm logical field describing how to treat NA values
(set to na.rm argument by default)
Sequence covariates must be ffTrack objects (or paths to ffTrack rds) and require an additional variables $signatures, which
will be used as input to fftab, and can have optional logical argument $grep to specify inexact matches (see fftab)
Interval covariates must be Granges (or paths to GRanges rds) or paths to bed files}

\item{out.path}{out.path to save variable to}

\item{out.path}{out.path to save variable to}
}
\value{
GRanges of input targets annotated with covariate statistics (+/- constrained to the subranges in optional argument covered)
}
\description{
Takes input of GRanges targets, an optional set of "covered" intervals, and an indefinite list of covariates which can be R objects
(GRanges, ffTrack, Rle) or file paths to .rds, .bw, .bed files, and an annotated target intervals GRanges with covariates computed
for each interval.   These target intervals can be further annotated with mutation counts and plugged into a generalized linear regression
(or other) model downstream.
}
\details{
There are three types of covariates: numeric, sequence, interval.  The covariates are computed as follows:
numeric covariates: the mean value
sequence covarites: fraction of bases satisfying $signature
interval covariates: fraction of bases overlapping feature
}
\author{
Marcin Imielinski
}
