% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishhook.R
\name{c.Cov_Arr}
\alias{c.Cov_Arr}
\title{title}
\usage{
\method{c}{Cov_Arr}(...)
}
\arguments{
\item{...}{A series of Covariates, note all objects must be of type Cov_Arr}
}
\value{
Cov_Arr object that can be passed directly into the FishHook object constructor that contains all of the Cov_Arr covariates
Passed in the ... param
}
\description{
Override the c operator for covariates so that you can merge them like a vector
}
\author{
Zoran Z. Gajic
}
