% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishhook.R
\docType{data}
\name{Cov_Arr}
\alias{Cov_Arr}
\title{Cov_Arr}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Cov_Arr
}
\arguments{
\item{...}{several Cov objects for packaging.}
}
\value{
Cov_Arr object that can be passed directly to the FishHook object constructor
}
\description{
Stores Covariates for passing to FishHook object constructor.Standard initialization involves calling c(Cov1,Cov2,Cov3). Cov_Arr serves to mask the underlieing list implemenations of Covariates in the FishHook Object. This class attempts to mimic a vector in terms of subsetting and in the future will add more vector like operations.
}
\author{
Zoran Z. Gajic
}
\keyword{datasets}
