% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishhook.R
\name{qq_pval}
\alias{qq_pval}
\title{qq plot given input p values}
\usage{
qq_pval(obs, highlight = c(), exp = NULL, lwd = 1, col = NULL,
  col.bg = "black", pch = 18, cex = 1, conf.lines = TRUE, max = NULL,
  max.x = NULL, max.y = NULL, qvalues = NULL, label = NULL,
  plotly = FALSE, annotations = list(), gradient = list(),
  titleText = "", subsample = NA, key = NULL, ...)
}
\arguments{
\item{obs}{vector of pvalues to plot, names of obs can be intepreted as labels}

\item{highlight}{vector optional arg specifying indices of data points to highlight (i.e. color red) (default = c())}

\item{exp}{info (default = NULL)}

\item{lwd}{integer, optional, specifying thickness of line fit to data (default = 1)}

\item{col}{info (default = NULL)}

\item{col.bg}{string}

\item{pch}{integer dot type for scatter plot}

\item{cex}{integer dot size for scatter plot}

\item{conf.lines}{logical, optional, whether to draw 95 percent confidence interval lines around x-y line}

\item{max}{numeric, optional, threshold to max the input p values}

\item{max.x}{info}

\item{max.y}{info}

\item{qvalues}{info}

\item{label}{character vector, optional specifying which data points to label (obs vector has to be named, for this to work)}

\item{plotly}{toggles between creating a pdf (FALSE) or an interactive html widget (TRUE)}

\item{annotations}{named list of vectors containing information to present as hover text (html widget), must be in same order as obs input}

\item{gradient}{named list that contains one vector that color codes points based on value, must bein same order as obs input}

\item{titleText}{title for plotly (html) graph only}
}
\description{
Explanation
}
\author{
Marcin Imielinski, Eran Hodis, Zoran Z. Gajic
}
