% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishhook.R
\name{score.targets}
\alias{score.targets}
\title{title}
\usage{
score.targets(targets, covariates = names(values(targets)), model = NULL,
  return.model = FALSE, nb = TRUE, verbose = TRUE, iter = 200,
  subsample = 1e+05, seed = 42, p.randomized = TRUE,
  classReturn = FALSE)
}
\arguments{
\item{targets}{annotated targets with fields $coverage, optional field, $count and additional numeric covariates}

\item{covariates}{info}

\item{model}{fit existing model --> covariates must be present (default = NULL)}

\item{return.model}{boolean info (default = FALSE)}

\item{nb}{boolean negative binomial, if false then use poisson}

\item{verbose}{boolean verbose flag (default = TRUE)}

\item{iter}{integer info (default = 200)}

\item{subsample}{interger info (default = 1e5)}

\item{seed}{integer (default = 42)}

\item{p.randomized}{boolean Flag info (default = TRUE)}

\item{classReturn}{boolean Flag info (default = FALSE)}
}
\value{
GRanges of scored results
}
\description{
Scores targets based on covariates using Gamma-Poisson model with coverage as constant
}
\author{
Marcin Imielinski
}
