% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishhook.R
\name{[.FishHook}
\alias{[.FishHook}
\title{title}
\usage{
\method{[}{FishHook}(obj, i, j, k, l)
}
\arguments{
\item{obj}{FishHook object This is the FishHookObject to be subset}

\item{i}{vector subset targets}

\item{j}{vector subset events}

\item{k}{vector subset covariats}

\item{l}{vector susbet eligible}
}
\value{
A new FishHook object that contains only the Covs within the given range
}
\description{
Overrides the subset operator x[] for use with FishHook to allow for vector like subsetting
}
\author{
Zoran Z. Gajic
}
